Making initialisation more configurable
A list of seats & options is now passed through when initalising
Popcorn.

I've also moved the click handling up out of the seat, to
prevent having to pass around the maximum number of seats.

Still need to figure out some of the data structure around storing the
seats. And figuring out if that's even needed.